requirements.txt
streamlit
langchain
langchain-core
langchain-google-genai
langchain-community
faiss-cpu
pypdf
tiktoken
python-dotenv